thegrouch/requirements.txt
# These requirements are only necessary when developing on Oscar.

# development
django-debug-toolbar>=2.2,<4.2
django-extensions>=2.2,<3.3
psycopg2-binary>=2.8,<2.10

Whoosh>=2.7,<2.8
django-environ>=0.4,<0.11
django-redis>=4.12,<5.3
pysolr>=3.9,<3.10
redis>=3.5,<4.6
requests>=2.25,<3
uWSGI>=2.0.19
whitenoise>=5.2,<6.5
sorl-thumbnail>=3.2

# Linting
pylint>=2.17.4
pylint-django>=2.5.3
black>=23.3.0

# Helpers
pyprof2calltree>=1.4,<1.5
ipdb>=0.13,<0.14
ipython>=7.12,<9

# Country data
pycountry

# Payment
stripe>=2.27.0

# Oscar API
django-oscar-api
